requirements.txt
aiohttp==3.8.3
jinja2==3.1.2
pyYAML==6.0
gino==1.0.1
alembic==1.8.1
SQLAlchemy~=1.3.24
cryptography==38.0.03
colorlog==6.7.0

google-auth==2.14.1
google-auth-oauthlib==0.7.1